algorithm/FGLM_files/requirements.txt
scipy~=1.7.3
numpy~=1.21.5
pandas~=1.3.5
fairlearn~=0.7.0
scikit-learn~=1.0.2
matplotlib~=3.5.1
pyyaml~=6.0
wget~=3.2
cvxpy~=1.2.1
dccp~=1.0.3
